pyproject.toml
[tool.poetry]
name = "google-workspace-agent"
version = "0.1.0"
description = "Intelligent agent for Google Workspace integration"
authors = ["Cline <cline@example.com>"]
readme = "README.md"
packages = [{include = "google_workspace_agent", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
google-auth = "*"
google-auth-oauthlib = "*"
google-auth-httplib2 = "*"
google-api-python-client = "*"
groq = "*"
pydantic = "*"
python-dotenv = "*"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
mypy = "^1.3.0"
black = "^23.3.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = true
ignore_missing_imports = true

[tool.black]
line-length = 100
target-version = ['py39']

[tool.flake8]
max-line-length = 100
extend-ignore = "E203,"
